Add missing kettle statuses
Thank you very much for this project! I bought this kettle a few days ago and was eager to integrate it with my home automation (which I did: https://gist.github.com/ghys/ea549402e88876e6daa3ac9aa403ec9f)

I have found a few statuses that are not in the KettleStatus enum.

```
ValueError: 'heating' is not a valid KettleStatus

ValueError: 'cooling_to_target' is not a valid KettleStatus
```
These are encountered when you enable the "boil before target" function and start.